Refactored project structure and requests
