+++ src/components/Card/StyledCardComponent.ts
