+++ src/components/CardDetails/CardDetails.tsx
