+++ src/components/CardDetails/index.ts
@@ -1,0 +1,1 @@
+export * from "./CardDetails";
