+++ src/components/CardsContainer/index.ts
@@ -1,0 +1,2 @@
+export * from "./CardsContainer";
+export * from "./StyledCardsContainer";
