+++ src/components/FormControl/FormControl.tsx
@@ -1,9 +1,9 @@
 import { Button } from "@material-ui/core";
 import styled from "styled-components";
-import { inputfFields, selects } from "../data";
-import { FormSelect } from "./FormSelect";
-import { InputFields } from "./InputFields";
-import { StyledSelectContainer } from "./StyledFormSelect";
+import { inputfFields, selects } from "../../data";
+import { FormInputFields } from "../FormInputFields";
+import { FormSelect } from "../FormSelect/index";
+import { StyledSelectContainer } from "../FormSelect/StyledFormSelect";
 
 const StyledButtonWrapper = styled.div`
   margin-top: 10px;
@@ -14,7 +14,7 @@ export const Form = (props: { onSubmit: (event: React.FormEvent) => void }) => {
   return (
     <>
       <form onSubmit={submit}>
-        <InputFields items={inputfFields} />
+        <FormInputFields items={inputfFields} />
         <StyledSelectContainer>
           {selects.map((select) => (
             <FormSelect key={select.id} name={select.name} id={select.id} label={select.label} items={select.items} />
