+++ src/components/FormSelect/index.ts
@@ -1,1 +1,2 @@
-export * from "../FormSelect";
+export * from "./FormSelect";
+export * from "./StyledFormSelect";
